Add files via upload
We have created a 2-dimensional turn-based maze game, 
 
The player can be moved using the arrow keys and we have implemented the Graphical user Interface to move Player on Mouse Click.
We have implemented BFS to figure out which enemy is present at the shortest distance by calculating distances of all enemies from the player. The nearest enemy only will move in the direction of the player.
BFS Vertices Placement on the Maze
 	1	2	3	4	5	6	7	8	 
9	10	 	 	11	12	 	 	13	14
 	15	16	17	18	19	20	21	22	 
23	24	25	26	 	 	27	28	29	30
31	32	33	34	35	36	37	38	39	40
41	 	 	 	43	44	 	 	 	45
46	47	48	49	50	51	52	53	54	55
56	57	58	 	 	59	 	60	 	 
 	 	61	62	63	64	 	65	66	67
68	69	70	 	71	72	73	74	 	75

We have Used Map and Vector to store and retrieve the nodes with their vertex location in the maze and used them to sort the distances of all Enemies from Player and find the Nearest Enemy.
Restricted Enemy and Player from colliding the wall. Player will be safe in Bush as Enemy is restricted from entering the Bush.
 


Player can fire after he collects the Set Of Arrows using Space Key.
Before: 						After:
                
Once the arrow hits the Enemies, the enemies will die and will Disappear.
 
And Finally, when the Player collect the Chest, He WINS!! 
 
On Contrary If the Enemy gets the Player before the Chest is Collected, GAME IS OVER!!